Update
With the help of the kick ass awesome facultee at QuantInsti, this
article has been updated to answer many of the questions some of the
readers had.

One of the biggest changes is that instead of only tracking the pnl on a
trade it now calculates the daily marked to market, or maybe a better
phrase will be the daily equity curve of market value of the portfolio.
(Previously I only tracked the profit or loss of a trade in the equity
curve.)